% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_legacy_api.R
\name{do_enrichment_tradeSeq}
\alias{do_enrichment_tradeSeq}
\title{do_enrichment_tradeSeq}
\usage{
do_enrichment_tradeSeq(
  tradeSeqRes,
  lineage_name,
  universe = NULL,
  species_db = org.Hs.eg.db,
  ont = "BP",
  do_kegg = TRUE,
  go_simplify = TRUE,
  simplify_cutoff = 0.7
)
}
\arguments{
\item{tradeSeqRes}{Data.frame with tradeSeq results (gene rownames).}

\item{lineage_name}{Character, e.g. "lineage1" (case-insensitive, will match column patterns).}

\item{universe}{Optional gene universe (SYMBOLs).}

\item{species_db}{OrgDb object (default org.Hs.eg.db).}

\item{ont}{GO ontology ("BP", "MF", "CC"), default "BP".}

\item{do_kegg}{Logical, whether to perform KEGG enrichment.}

\item{go_simplify}{Logical, whether GO terms should be simplified to reduce redundancy using clusterProfiler::simplify. Default is TRUE}

\item{simplify_cutoff}{similarity threshold}
}
\value{
List with enrichment results for up/down genes (GO/KEGG).
}
\description{
Wrapper to perform GO and KEGG enrichment for up- and down-regulated genes from tradeSeq results for a specified lineage.
}
