% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_enrichment.R
\name{enrich_go_clusters}
\alias{enrich_go_clusters}
\title{Calculate GO terms separately for each gene set}
\usage{
enrich_go_clusters(gene_sets, organism_db, ontology = "BP")
}
\arguments{
\item{gene_sets}{Named list of Entrez ID vectors.}

\item{organism_db}{OrgDb object.}

\item{ontology}{GO ontology.}
}
\value{
Named list of enrichment objects.
}
\description{
Calculate GO terms separately for each gene set
}
