% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_enrichment.R
\name{enrich_go_terms}
\alias{enrich_go_terms}
\title{Calculate GO enrichment without plotting or writing files}
\usage{
enrich_go_terms(genes, organism_db, ontology = "BP", p_cutoff = 0.05)
}
\arguments{
\item{genes}{Entrez IDs.}

\item{organism_db}{OrgDb object.}

\item{ontology}{GO ontology.}

\item{p_cutoff}{Adjusted significance cutoff.}
}
\value{
\code{clusterProfiler} enrichment object.
}
\description{
Calculate GO enrichment without plotting or writing files
}
