% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_export.R
\name{export_excel}
\alias{export_excel}
\title{Export one or more data frames to Excel}
\usage{
export_excel(data, file = "results.xlsx", rownames = FALSE, styled = TRUE)
}
\arguments{
\item{data}{Data frame or named list of data frames.}

\item{file}{Destination filename.}

\item{rownames}{Include row names.}

\item{styled}{Style the header, shade alternate rows and freeze the first row.}
}
\value{
Filename, invisibly. Parent directories are created if necessary.
}
\description{
Export one or more data frames to Excel
}
