% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sample_sheet.R
\name{format_sample_sheet}
\alias{format_sample_sheet}
\title{Save workspace and significant sig.res table from NGS analysis}
\usage{
format_sample_sheet(sample.sheet)
}
\arguments{
\item{sample.sheet}{csv sample sheet from the Illumina NextSeq 1000 sequencing platform.}
}
\description{
Save workspace and significant sig.res table from NGS analysis
}
