% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_formula.R
\name{make_design_formula}
\alias{make_design_formula}
\title{Construct design formula out of 1 or 2 factors}
\usage{
make_design_formula(fct1, fct2 = NULL)
}
\arguments{
\item{fct1}{Character string. Name of the first factor.}

\item{fct2}{Character string. Name of an optional second factor.}
}
\description{
Construct design formula out of 1 or 2 factors
}
