% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_save.R
\name{obj_save}
\alias{obj_save}
\title{Save r objects as individual rds files.}
\usage{
obj_save(..., folder)
}
\arguments{
\item{...}{object names to be saved individually as .rds files.}

\item{folder}{folder where objects should be stored.}
}
\description{
Save r objects as individual rds files.
}
