% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_legacy_api.R
\name{pca_plot}
\alias{pca_plot}
\title{Function to plot the first 2 principal components of a rlog or vst transformed count matrix.}
\usage{
pca_plot(
  data,
  pcsToUse,
  title = "",
  subtitle = "",
  type = "VST",
  colors = NULL,
  shapes = NULL,
  pointSize = 3,
  textSize = 12,
  labelled = FALSE,
  intgroup
)
}
\arguments{
\item{data}{Transformed count matrix.}

\item{pcsToUse}{numeric of length 2, which PCs to plot}

\item{title}{plot title.}

\item{subtitle}{plot subtitle.}

\item{type}{Specify type of transformation which was used for generation of count matrix.}

\item{colors}{character vector with colors used in scale_color_manual. Default is npg colors from ggsci package.}

\item{shapes}{character vector with shapes. Default are ggplot shapes if no shapes are supplied.}

\item{pointSize}{Size of points.}

\item{textSize}{Size of plot text.}

\item{labelled}{logical indicating whether points or text labels should be plotted. Default is FALSE to plot points.}

\item{intgroup}{interesting groups: a character vector of
names in \code{colData(x)} to use for grouping}
}
\description{
Output is a ggplot object showing PC1 and PC2 for a RNASeq experiment. For a 1 factor design, points are colored by this group,
for a 2 factor design the first factor is represented by colors the second factor by shape.
}
