% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_legacy_api.R
\name{plot_go}
\alias{plot_go}
\title{Function to plot GO clusters}
\usage{
plot_go(
  genes,
  showCategory = 20,
  sim.thres = 0.7,
  sym.colors = FALSE,
  return.res = FALSE,
  font.size = 12,
  pvalueCutoff = 0.05,
  path
)
}
\arguments{
\item{genes}{Vector containing ENTREZ IDs of differentially expressed genes.}

\item{showCategory}{A number or a list of terms. If it is a number, the first n terms will be displayed. If it is a list of terms, the selected terms will be displayed.}

\item{sim.thres}{similarity threshold (0-1). Some guidance: Large (allowed similarity=0.9), Medium (0.7), Small (0.5), Tiny (0.4) Defaults to Medium (0.7)}

\item{sym.colors}{Logical indicating whether colors distribution should be symmetrical. Default is FALSE.}

\item{return.res}{Logical indicating whether results should be returned. If TRUE original enrichGO results and reduced terms will be stored as list object.}

\item{font.size}{Font size.}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report}

\item{path}{Path were results should be stored. Can be an absolute path or relative path based on the working directory.}
}
\description{
This function uses a gene list of ENTREZ IDs as input for clusterProfiler's function \link[clusterProfiler]{compareCluster} with the parameters \emph{fun = enrichGO},
\emph{ont = "BP"} and \emph{OrgDb = org.Hs.eg.db}. The package \emph{org.Hs.eg.db} is required for this function to work properly.
}
