% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_legacy_api.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Function to plot and save}
\usage{
plot_wordcloud(text)
}
\arguments{
\item{text}{Character vector containing a list of pathway strings.}
}
\description{
Text is searched for common terms and a wordcloud is generated based on frequency.
}
