% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_input.R
\name{print_input}
\alias{print_input}
\title{Function to print input thresholds for the DESeq pipeline.}
\usage{
print_input(
  SummarizedExp,
  DESeqObj,
  pThres = 0.05,
  lfcThres = 1,
  minimumCount = 10,
  smallestGroupSize,
  contrast,
  fontSize = 9
)
}
\arguments{
\item{SummarizedExp}{SummarizedExperiment object.}

\item{DESeqObj}{DESeq object.}

\item{pThres}{p-Value threshold.}

\item{lfcThres}{log fold-change threshold}

\item{minimumCount}{minimum read count for a gene to be included. The threshold works together with the smallestGroupSize
argument. For example a minimumCount = 10 and smallestGroupSize = 3 would eliminate all genes who do not reach a count of 10
in at least 3 samples.}

\item{smallestGroupSize}{sample size of the smallest group. For more details see argument minimumCount.}

\item{contrast}{design formula as specified to obtain the DESeq object.}

\item{fontSize}{size of table font.}
}
\description{
Output is a flextable object with prespecified styling. Needs a SummarizedExperiment object and a DESeq Object as well
as significant and log fold-change thresholds and the minimum read count size as well as the minimum group Size. Additionally
the contrast formula needs to be specified.
}
