% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_reads.R
\name{print_reads}
\alias{print_reads}
\title{Function to print reads per sample}
\usage{
print_reads(
  readsRaw,
  readsFiltered,
  minimumCount = 10,
  smallestGroupSize,
  fontSize = 8
)
}
\arguments{
\item{readsRaw}{named vector including the total number of reads per sample.}

\item{readsFiltered}{named vector including the total number of reads per sample after applying filtering (e.g. removing lowly expressed transcripts).}

\item{minimumCount}{minimum read count for a gene to be included. The threshold works together with the smallestGroupSize
argument. For example a minimumCount = 10 and smallestGroupSize = 3 would eliminate all genes who do not reach a count of 10
in at least 3 samples.}

\item{smallestGroupSize}{sample size of the smallest group. For more details see argument minimumCount.}

\item{fontSize}{size of table font.}
}
\description{
Output is a flextable object with prespecified styling.
}
