% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check.R
\name{sanity_check}
\alias{sanity_check}
\title{Sanity check for NGS data prepared with DESeq}
\usage{
sanity_check(
  rdata,
  deseq.obj = "dds",
  rlog.cm = "rld",
  all.res = "allRes",
  res.df = "res"
)
}
\arguments{
\item{rdata}{rdata object containing DESeq results which will be used for the downstream analysis}

\item{deseq.obj}{DESeq2 object}

\item{rlog.cm}{rlog transformed count matrix}

\item{all.res}{annotated results data.frame}

\item{res.df}{results data.frame}
}
\description{
This function checks whether the dds, rld, res and allRes object all have the same lengths or if discrepancies are
present.
}
