% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraster.R
\name{contraster}
\alias{contraster}
\title{Function to define complex contrasts in DESeq results function.}
\usage{
contraster(dds, group1, group2, weighted = F)
}
\arguments{
\item{dds}{DESeq object containing colData and design}

\item{group1}{list of character vectors each with 2 or more items}

\item{group2}{list of character vectors each with 2 or more items}

\item{weighted}{logical indicating whether weighted contrasts should be applied. Default is FALSE.}
}
\description{
Function is taken from https://www.atakanekiz.com/technical/a-guide-to-designs-and-contrasts-in-DESeq2/ to allow complex
contrasts including difference of differences and individual comparisons
}
