% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goseq_to_revigo.R
\name{goseq_to_revigo}
\alias{goseq_to_revigo}
\title{Perform goseq analysis and reduce terms based on similarity threshold}
\usage{
goseq_to_revigo(
  gene.list,
  genome,
  identifier = "ensGene",
  ont = "BP",
  sim.thres = 0.7,
  pval.thres = 0.05
)
}
\arguments{
\item{gene.list}{named gene list containing 1 (gene of interest) or 0 (other gene). Names can be either Ensemble IDs
or ENTREZ IDs. Based on the name the identifier has to be set to either "ensGene" or "knownGene" for Ensemble and Entrez IDs, respectively.}

\item{genome}{reference genome. For further details see \code{goseq} function \code{\link[=nullp]{nullp()}}.}

\item{identifier}{gene identifier as specified in gene.list.}

\item{ont}{GO ontology. Either "BP" for "biological process", "MF" for "molecular function" or "CC" for cellular compartment. Default is "BP".}

\item{sim.thres}{similarity threshold passed on to the \code{rrvgo} function \code{\link[=reduceSimMatrix]{reduceSimMatrix()}}.}

\item{pval.thres}{threshold to filter enriched pathways.}
}
\description{
This function takes a named gene list with Gene IDs as names and 1 or 0 for significantly or unchanged genes, respectively, as input.
It is possible to analyze any gene set by assigning 1 to the genes of interest and 0 for other genes.
}
