% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dispersion_genes.R
\name{high_dispersion_genes}
\alias{high_dispersion_genes}
\title{Extract high dispersion genes}
\usage{
high_dispersion_genes(
  dds.obj,
  norm.counts,
  res,
  min.count = 10,
  n.min = dim(dds.obj)[2]/2,
  disp.thres = 1
)
}
\arguments{
\item{dds.obj}{DESeq2 object containg differential expression data}

\item{norm.counts}{normalized count data (e.g. rlog or vst transformed count data).}

\item{res}{results table containing gene IDs as ENSEMBL ID as well as gene names stored as "SYMBOL".}

\item{min.count}{remove low expressed genes which would interfere with the calculation of dispersion estimates.}

\item{n.min}{the smallest sample size of the experiment which refers to the number of samples which need a count above the threshold set in
min.count to be considered.}

\item{disp.thres}{dispersion threshold. All genes higher than the specified number will be extracted.}
}
\description{
Extract high dispersion genes
}
