% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_signif_genes.R
\name{list_signif_genes}
\alias{list_signif_genes}
\title{Function to list ENTREZ IDs of significantly changed genes}
\usage{
list_signif_genes(
  list,
  p.threshold = 0.05,
  lfc.threshold = 0,
  direction = c("greater", "lesser")
)
}
\arguments{
\item{list}{List containing results of DE analysis generated with \link[DESeq2]{results} function.}

\item{p.threshold}{P-value threshold used to define differentially expressed genes}

\item{lfc.threshold}{Log-fold change threshold used to define differentially expressed genes.}

\item{direction}{Character specifying whether ENTREZ IDs of upregulated ("greater") or downregulated ("lesser") genes should be extracted}
}
\description{
This function lists significantly up or downregulated genes as ENTREZ IDs for later use in clusterProfiler's function \link[clusterProfiler]{compareCluster}.
}
