% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{Function to plot the first 2 principal components of a rlog or vst transformed count matrix.}
\usage{
pca_plot(
  data,
  percentVar,
  type,
  pointSize = 3,
  textSize = 12,
  labelled = FALSE,
  fct1,
  fct2 = NULL
)
}
\arguments{
\item{data}{Transformed count matrix.}

\item{percentVar}{Explained variance by PC1 and PC2}

\item{type}{Specify type of transformation which was used for generation of count matrix.}

\item{pointSize}{Size of points.}

\item{textSize}{Size of plot text.}

\item{labelled}{logical indicating whether points or text labels should be plotted. Default is FALSE to plot points.}

\item{fct1}{main factor of interest in the design formula}

\item{fct2}{covariate which is adjusted for if included in the design formula}
}
\description{
Output is a ggplot object showing PC1 and PC2 for a RNASeq experiment. For a 1 factor design, points are colored by this group,
for a 2 factor design the first factor is represented by colors the second factor by shape.
}
