% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_clusters.R
\name{plot_go_clusters}
\alias{plot_go_clusters}
\title{Function to plot and save}
\usage{
plot_go_clusters(gene.list, pvalueCutoff = 0.05, path)
}
\arguments{
\item{gene.list}{List containing ENTREZ IDs of enriched genes. List should be generated with the list_signif_genes() function from the ekbSeq package.}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report}

\item{path}{Path were results should be stored. Can be an absolute path or relative path based on the working directory.}
}
\description{
This function uses a gene list of ENTREZ IDs as input for clusterProfiler's function \link[clusterProfiler]{compareCluster} with the parameters \emph{fun = enrichGO},
\emph{ont = "BP} and \emph{OrgDb = org.Hs.eg.db}. The package \emph{org.Hs.eg.db} is required for this function to work properly.
}
