% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_res.R
\name{save_res}
\alias{save_res}
\title{Save workspace and significant results table from NGS analysis}
\usage{
save_res(save = TRUE, results = NULL, path = NULL)
}
\arguments{
\item{save}{logical indicating whether workspace and significant results table should be saved. Default is TRUE.}

\item{results}{data.frame containing results with differentially expressed genes.}

\item{path}{path where data should be saved. Must end with "<\if{html}{\out{<filename>}}>.csv".}
}
\description{
Save workspace and significant results table from NGS analysis
}
