% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_res.R
\name{save_res}
\alias{save_res}
\title{Save workspace and significant sig.res table from NGS analysis}
\usage{
save_res(
  save = TRUE,
  all.res = NULL,
  sig.res = NULL,
  pathway.res = NULL,
  results.name = NULL,
  rdata.name,
  rm.batch = NULL
)
}
\arguments{
\item{save}{logical indicating whether workspace and significant results table should be saved. Default is TRUE.}

\item{all.res}{data.frame containing results with expression data.}

\item{sig.res}{data.frame containing results with differentially expressed genes.}

\item{pathway.res}{data.frame containing results of pathfindR using wiki pathways.}

\item{results.name}{name of csv file where differential expression results are stored.}

\item{rdata.name}{name of the exported rdata file.}

\item{rm.batch}{second factor of a design formula. If present the batch effect introduced by this covariate is removed and the corresponding
files are saved with the suffix "adjusted".}
}
\description{
Save workspace and significant sig.res table from NGS analysis
}
